+++ CHANGELOG.md
@@ -50,3 +50,4 @@
 ### 1.2.5 (2023-12-14)
 
 - Replaced deprecated ethereumjs-tx lib to @ethereumjs/tx and @ethereumjs/common for transaction signing.
+- Changed license to MIT.